+++ trace-forward/src/Trace/Forward/Forwarding.hs
@@ -111,8 +111,9 @@ initForwardingDelayed iomgr config magic ekgStore tracerSocketMode = liftIO $ do
       Nothing -> EKGF.LocalPipe ""
       Just (LocalPipe str, _mode) -> EKGF.LocalPipe str
       Just (RemoteSocket host port, _mode) -> EKGF.RemoteSocket host port
-  connSize = tofConnQueueSize config
-  disconnSize = tofDisconnQueueSize config
+  -- Having two queue sizes was removed from the "trace-foirward" package.
+  -- It will be removed from `TraceOptionForwarder` and the user configuration.
+  queueSize = max (tofConnQueueSize config) (tofDisconnQueueSize config)
   verbosity = tofVerbosity config
   maxReconnectDelay = tofMaxReconnectDelay config
 
@@ -129,9 +130,8 @@ initForwardingDelayed iomgr config magic ekgStore tracerSocketMode = liftIO $ do
   tfConfig :: TF.ForwarderConfiguration TraceObject
   tfConfig =
     TF.ForwarderConfiguration
-      { TF.forwarderTracer       = mkTracer verbosity
-      , TF.disconnectedQueueSize = disconnSize
-      , TF.connectedQueueSize    = connSize
+      { TF.forwarderTracer = mkTracer verbosity
+      , TF.queueSize       = queueSize
       }
 
   dpfConfig :: DPF.ForwarderConfiguration