+++ trace-forward/src/Trace/Forward/Utils/ForwardSink.hs
@@ -1,15 +1,11 @@
-{-# LANGUAGE DataKinds #-}
-{-# LANGUAGE GADTs #-}
-
 module Trace.Forward.Utils.ForwardSink
   ( ForwardSink (..)
   ) where
 
-import           Control.Concurrent.STM.TBQueue
+import Control.Concurrent.STM.TBQueue
 
 data ForwardSink lo = ForwardSink
   { forwardQueue     :: !(TBQueue lo)
-  , disconnectedSize :: !Word
-  , connectedSize    :: !Word
   , overflowCallback :: !([lo] -> IO ())
   }
+