+++ trace-forward/CHANGELOG.md
@@ -1,6 +1,11 @@
 # ChangeLog
 
+## NEXT - Nov 2025
+
+* Replaced both `disconnectedQueueSize` and `connectedQueueSize` with `queueSize` (See #6361 for details).
+
 ## 2.3.1 - Oct 2025
+
 * Updated to `ekg-forward-1.0`, `ouroboros-network-framework-0.19.2` and `typed-protocols-1.0`.
 
 ## 2.3.0 - Jul 2025